Application created
In this app you can manage your library by adding and deleting books from it as well as accept and deny requests from users, who want to borrow books. Also you can login as a user and borrow any book for reading (only if the librarian allows you to do it).